headless_lms_chatbot

Module azure_search_indexer

Structs

IndexerError 🔒

IndexerStatusResponse 🔒

IndexerWarning 🔒

LastResult 🔒

Constants

API_VERSION 🔒

Functions

check_search_indexer_status

Checks if the search indexer exists, is not running, and the last execution is not in progress. Also prints any errors and warnings from the last execution in a nicely formatted manner.

create_search_indexer

does_search_indexer_exist

run_search_indexer_now