headless_lms_chatbot::chatbot_tools::custom_tools

Module document_lookup

Structs

DocumentLookupArguments

DocumentLookupState

Functions

deserialize_to_optional_uuid_and_errors_to_none 🔒

Deserializes an optional string field and parses it into an Uuid, doing this optionally without failure. If an error occurs, a None is returned. This is desired for the page_id field in DocumentLookupArguments, which is generated by an LLM and can be None or a non-Uuid string in some cases. If any errors should occur, they are emitted in ChatbotTools’s from_db_and_arguments.

shorten_page_content 🔒

Type Aliases

DocumentLookupTool