headless_lms_chatbot::chatbot_tools::custom_tools::document_lookup

Function deserialize_to_optional_uuid_and_errors_to_none

fn deserialize_to_optional_uuid_and_errors_to_none<'de, D>(
    deserializer: D,
) -> Result<Option<Uuid>, D::Error>
where
    D: Deserializer<'de>,

Deserializes an optional string field and parses it into an Uuid, doing this optionally without failure. If an error occurs, a None is returned. This is desired for the page_id field in DocumentLookupArguments, which is generated by an LLM and can be None or a non-Uuid string in some cases. If any errors should occur, they are emitted in ChatbotTools’s from_db_and_arguments.