headless_lms_chatbot

Module prelude

Re-exports

pub use headless_lms_models as models;

Macros

error

Constructs an event at the error level.

info

Constructs an event at the info level.

warn

Constructs an event at the warn level.

Structs

ApplicationConfiguration

PgConnection

A connection to a PostgreSQL database.

Uuid

A Universally Unique Identifier (UUID).

Statics

REQWEST_CLIENT

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Derive Macros

Deserialize

Serialize