headless_lms_chatbot::prelude::chrono::prelude

Enum Month

pub enum Month {
    January = 0,
    February = 1,
    March = 2,
    April = 3,
    May = 4,
    June = 5,
    July = 6,
    August = 7,
    September = 8,
    October = 9,
    November = 10,
    December = 11,
}

The month of the year.

This enum is just a convenience implementation. The month in dates created by DateLike objects does not return this enum.

It is possible to convert from a date to a month independently

use chrono::prelude::*;
let date = Utc.with_ymd_and_hms(2019, 10, 28, 9, 10, 11).unwrap();
// `2019-10-28T09:10:11Z`
let month = Month::try_from(u8::try_from(date.month()).unwrap()).ok();
assert_eq!(month, Some(Month::October))

Or from a Month to an integer usable by dates

let month = Month::January;
let dt = Utc.with_ymd_and_hms(2019, month.number_from_month(), 28, 9, 10, 11).unwrap();
assert_eq!((dt.year(), dt.month(), dt.day()), (2019, 1, 28));

Allows mapping from and to month, from 1-January to 12-December. Can be Serialized/Deserialized with serde

Variants

January = 0

January

February = 1

February

March = 2

March

April = 3

April

May = 4

May

June = 5

June

July = 6

July

August = 7

August

September = 8

September

October = 9

October

November = 10

November

December = 11

December

Implementations

impl Month

pub const fn succ(&self) -> Month

The next month.

m:	January	February	...	December

m.succ():	February	March	...	January

pub const fn pred(&self) -> Month

The previous month.

m:	January	February	...	December

m.pred():	December	January	...	November

pub const fn number_from_month(&self) -> u32

Returns a month-of-year number starting from January = 1.

m:	January	February	...	December

m.number_from_month():	1	2	...	12

pub const fn name(&self) -> &'static str

Get the name of the month

use chrono::Month;

assert_eq!(Month::January.name(), "January")

pub fn num_days(&self, year: i32) -> Option<u8>

Get the length in days of the month

Yields None if year is out of range for NaiveDate.