headless_lms_chatbot::prelude::tokio::runtime

Struct Builder

pub struct Builder {

    kind: Kind,
    name: Option<String>,
    enable_io: bool,
    nevents: usize,
    enable_time: bool,
    start_paused: bool,
    worker_threads: Option<usize>,
    max_blocking_threads: usize,
    pub(super) thread_name: Arc<dyn Fn() -> String + Sync + Send>,
    pub(super) thread_stack_size: Option<usize>,
    pub(super) after_start: Option<Arc<dyn Fn() + Sync + Send>>,
    pub(super) before_stop: Option<Arc<dyn Fn() + Sync + Send>>,
    pub(super) before_park: Option<Arc<dyn Fn() + Sync + Send>>,
    pub(super) after_unpark: Option<Arc<dyn Fn() + Sync + Send>>,
    pub(super) before_spawn: Option<Arc<dyn Fn(&TaskMeta<'_>) + Sync + Send>>,
    pub(super) after_termination: Option<Arc<dyn Fn(&TaskMeta<'_>) + Sync + Send>>,
    pub(super) keep_alive: Option<Duration>,
    pub(super) global_queue_interval: Option<u32>,
    pub(super) event_interval: u32,
    pub(super) disable_lifo_slot: bool,
    pub(super) seed_generator: RngSeedGenerator,
    pub(super) metrics_poll_count_histogram_enable: bool,
    pub(super) metrics_poll_count_histogram: HistogramBuilder,
    timer_flavor: TimerFlavor,
    enable_eager_driver_handoff: bool,

}

Builds Tokio Runtime with custom configuration values.

Methods can be chained in order to set the configuration values. The Runtime is constructed by calling build.

New instances of Builder are obtained via Builder::new_multi_thread or Builder::new_current_thread.

See function level documentation for details on the various configuration settings.

Examples

use tokio::runtime::Builder;

fn main() {
    // build runtime
    let runtime = Builder::new_multi_thread()
        .worker_threads(4)
        .thread_name("my-custom-name")
        .thread_stack_size(3 * 1024 * 1024)
        .build()
        .unwrap();

    // use runtime ...
}

Fields
kind: Kindname: Option<String>enable_io: boolnevents: usizeenable_time: boolstart_paused: boolworker_threads: Option<usize>max_blocking_threads: usizethread_name: Arc<dyn Fn() -> String + Sync + Send>thread_stack_size: Option<usize>after_start: Option<Arc<dyn Fn() + Sync + Send>>before_stop: Option<Arc<dyn Fn() + Sync + Send>>before_park: Option<Arc<dyn Fn() + Sync + Send>>after_unpark: Option<Arc<dyn Fn() + Sync + Send>>before_spawn: Option<Arc<dyn Fn(&TaskMeta<'_>) + Sync + Send>>after_termination: Option<Arc<dyn Fn(&TaskMeta<'_>) + Sync + Send>>keep_alive: Option<Duration>global_queue_interval: Option<u32>event_interval: u32disable_lifo_slot: boolseed_generator: RngSeedGeneratormetrics_poll_count_histogram_enable: boolmetrics_poll_count_histogram: HistogramBuildertimer_flavor: TimerFlavorenable_eager_driver_handoff: bool
Implementations

impl Builder

pub fn new_current_thread() -> Builder

Returns a new builder with the current thread scheduler selected.

Configuration methods can be chained on the return value.

To spawn non-Send tasks on the resulting runtime, combine it with a LocalSet, or call build_local to create a LocalRuntime.

pub fn new_multi_thread() -> Builder

Returns a new builder with the multi thread scheduler selected.

Configuration methods can be chained on the return value.

pub fn enable_all(&mut self) -> &mut Builder

Enables both I/O and time drivers.

Doing this is a shorthand for calling enable_io and enable_time individually. If additional components are added to Tokio in the future, enable_all will include these future components.

Examples

use tokio::runtime;

let rt = runtime::Builder::new_multi_thread()
    .enable_all()
    .build()
    .unwrap();

pub fn worker_threads(&mut self, val: usize) -> &mut Builder

Sets the number of worker threads the Runtime will use.

This can be any number above 0 though it is advised to keep this value on the smaller side.

This will override the value read from environment variable TOKIO_WORKER_THREADS.

Default

The default value is the number of cores available to the system.

When using the current_thread runtime this method has no effect.

Examples

Multi threaded runtime with 4 threads

use tokio::runtime;

// This will spawn a work-stealing runtime with 4 worker threads.
let rt = runtime::Builder::new_multi_thread()
    .worker_threads(4)
    .build()
    .unwrap();

rt.spawn(async move {});

Current thread runtime (will only run on the current thread via Runtime::block_on)

use tokio::runtime;

// Create a runtime that _must_ be driven from a call
// to `Runtime::block_on`.
let rt = runtime::Builder::new_current_thread()
    .build()
    .unwrap();

// This will run the runtime and future on the current thread
rt.block_on(async move {});

Panics

This will panic if val is not larger than 0.

pub fn max_blocking_threads(&mut self, val: usize) -> &mut Builder

Specifies the limit for additional threads spawned by the Runtime.

These threads are used for blocking operations like tasks spawned through spawn_blocking, this includes but is not limited to:

fs operations

dns resolution through ToSocketAddrs

writing to Stdout or Stderr

reading from Stdin

Unlike the worker_threads, they are not always active and will exit if left idle for too long. You can change this timeout duration with thread_keep_alive.

It’s recommended to not set this limit too low in order to avoid hanging on operations requiring spawn_blocking.

The default value is 512.

Queue Behavior

When a blocking task is submitted, it will be inserted into a queue. If available, one of the idle threads will be notified to run the task. Otherwise, if the threshold set by this method has not been reached, a new thread will be spawned. If no idle thread is available and no more threads are allowed to be spawned, the task will remain in the queue until one of the busy threads pick it up. Note that since the queue does not apply any backpressure, it could potentially grow unbounded.

Panics

This will panic if val is not larger than 0.

Upgrading from 0.x

In old versions max_threads limited both blocking and worker threads, but the current max_blocking_threads does not include async worker threads in the count.

pub fn thread_name(&mut self, val: impl Into<String>) -> &mut Builder

Sets name of threads spawned by the Runtime’s thread pool.

The default name is “tokio-rt-worker”.

Examples

let rt = runtime::Builder::new_multi_thread()
    .thread_name("my-pool")
    .build();

pub fn name(&mut self, val: impl Into<String>) -> &mut Builder

Sets the name of the runtime.

Examples

let rt = runtime::Builder::new_multi_thread()
    .name("my-runtime")
    .build();

Panics

This function will panic if an empty value is passed as an argument.

pub fn thread_name_fn<F>(&mut self, f: F) -> &mut Builder
where F: Fn() -> String + Send + Sync + 'static,

Sets a function used to generate the name of threads spawned by the Runtime’s thread pool.

The default name fn is || "tokio-rt-worker".into().

Examples

let rt = runtime::Builder::new_multi_thread()
    .thread_name_fn(|| {
       static ATOMIC_ID: AtomicUsize = AtomicUsize::new(0);
       let id = ATOMIC_ID.fetch_add(1, Ordering::SeqCst);
       format!("my-pool-{}", id)
    })
    .build();

pub fn thread_stack_size(&mut self, val: usize) -> &mut Builder

Sets the stack size (in bytes) for worker threads.

The actual stack size may be greater than this value if the platform specifies minimal stack size.

The default stack size for spawned threads is 2 MiB, though this particular stack size is subject to change in the future.

Examples

let rt = runtime::Builder::new_multi_thread()
    .thread_stack_size(32 * 1024)
    .build();

pub fn on_thread_start<F>(&mut self, f: F) -> &mut Builder
where F: Fn() + Send + Sync + 'static,

Executes function f after each thread is started but before it starts doing work.

This is intended for bookkeeping and monitoring use cases.

Examples

let runtime = runtime::Builder::new_multi_thread()
    .on_thread_start(|| {
        println!("thread started");
    })
    .build();

pub fn on_thread_stop<F>(&mut self, f: F) -> &mut Builder
where F: Fn() + Send + Sync + 'static,

Executes function f before each thread stops.

This is intended for bookkeeping and monitoring use cases.

Examples

{
let runtime = runtime::Builder::new_multi_thread()
    .on_thread_stop(|| {
        println!("thread stopping");
    })
    .build();

pub fn on_thread_park<F>(&mut self, f: F) -> &mut Builder
where F: Fn() + Send + Sync + 'static,

Executes function f just before a thread is parked (goes idle). f is called within the Tokio context, so functions like tokio::spawn can be called, and may result in this thread being unparked immediately.

This can be used to start work only when the executor is idle, or for bookkeeping and monitoring purposes.

Note: There can only be one park callback for a runtime; calling this function more than once replaces the last callback defined, rather than adding to it.

Examples

Multithreaded executor

let once = AtomicBool::new(true);
let barrier = Arc::new(Barrier::new(2));

let runtime = runtime::Builder::new_multi_thread()
    .worker_threads(1)
    .on_thread_park({
        let barrier = barrier.clone();
        move || {
            let barrier = barrier.clone();
            if once.swap(false, Ordering::Relaxed) {
                tokio::spawn(async move { barrier.wait().await; });
           }
        }
    })
    .build()
    .unwrap();

runtime.block_on(async {
   barrier.wait().await;
})

Current thread executor

let once = AtomicBool::new(true);
let barrier = Arc::new(Barrier::new(2));

let runtime = runtime::Builder::new_current_thread()
    .on_thread_park({
        let barrier = barrier.clone();
        move || {
            let barrier = barrier.clone();
            if once.swap(false, Ordering::Relaxed) {
                tokio::spawn(async move { barrier.wait().await; });
           }
        }
    })
    .build()
    .unwrap();

runtime.block_on(async {
   barrier.wait().await;
})

pub fn on_thread_unpark<F>(&mut self, f: F) -> &mut Builder
where F: Fn() + Send + Sync + 'static,

Executes function f just after a thread unparks (starts executing tasks).

This is intended for bookkeeping and monitoring use cases; note that work in this callback will increase latencies when the application has allowed one or more runtime threads to go idle.

Note: There can only be one unpark callback for a runtime; calling this function more than once replaces the last callback defined, rather than adding to it.

Examples

let runtime = runtime::Builder::new_multi_thread()
    .on_thread_unpark(|| {
        println!("thread unparking");
    })
    .build();

runtime.unwrap().block_on(async {
   tokio::task::yield_now().await;
   println!("Hello from Tokio!");
})

pub fn build(&mut self) -> Result<Runtime, Error>

Creates the configured Runtime.

The returned Runtime instance is ready to spawn tasks.

Examples

use tokio::runtime::Builder;

let rt  = Builder::new_multi_thread().build().unwrap();

rt.block_on(async {
    println!("Hello from the Tokio runtime");
});

pub fn build_local( &mut self, options: LocalOptions, ) -> Result<LocalRuntime, Error>

Creates the configured LocalRuntime.

The returned LocalRuntime instance is ready to spawn tasks.

Panics

This will panic if the runtime is configured with new_multi_thread().

Examples

use tokio::runtime::{Builder, LocalOptions};

let rt = Builder::new_current_thread()
    .build_local(LocalOptions::default())
    .unwrap();

rt.spawn_local(async {
    println!("Hello from the Tokio runtime");
});

pub fn thread_keep_alive(&mut self, duration: Duration) -> &mut Builder

Sets a custom timeout for a thread in the blocking pool.

By default, the timeout for a thread is set to 10 seconds. This can be overridden using .thread_keep_alive().

Example

let rt = runtime::Builder::new_multi_thread()
    .thread_keep_alive(Duration::from_millis(100))
    .build();

pub fn global_queue_interval(&mut self, val: u32) -> &mut Builder

Sets the number of scheduler ticks after which the scheduler will poll the global task queue.

A scheduler “tick” roughly corresponds to one poll invocation on a task.

By default the global queue interval is 31 for the current-thread scheduler. Please see the module documentation for the default behavior of the multi-thread scheduler.

Schedulers have a local queue of already-claimed tasks, and a global queue of incoming tasks. Setting the interval to a smaller value increases the fairness of the scheduler, at the cost of more synchronization overhead. That can be beneficial for prioritizing getting started on new work, especially if tasks frequently yield rather than complete or await on further I/O. Setting the interval to 1 will prioritize the global queue and tasks from the local queue will be executed only if the global queue is empty. Conversely, a higher value prioritizes existing work, and is a good choice when most tasks quickly complete polling.

Panics

This function will panic if 0 is passed as an argument.

Examples

let rt = runtime::Builder::new_multi_thread()
    .global_queue_interval(31)
    .build();

pub fn event_interval(&mut self, val: u32) -> &mut Builder

Sets the number of scheduler ticks after which the scheduler will poll for external events (timers, I/O, and so on).

A scheduler “tick” roughly corresponds to one poll invocation on a task.

By default, the event interval is 61 for all scheduler types.

Setting the event interval determines the effective “priority” of delivering these external events (which may wake up additional tasks), compared to executing tasks that are currently ready to run. A smaller value is useful when tasks frequently spend a long time in polling, or infrequently yield, which can result in overly long delays picking up I/O events. Conversely, picking up new events requires extra synchronization and syscall overhead, so if tasks generally complete their polling quickly, a higher event interval will minimize that overhead while still keeping the scheduler responsive to events.

Panics

This function will panic if 0 is passed as an argument.

Examples

let rt = runtime::Builder::new_multi_thread()
    .event_interval(31)
    .build();

impl Builder

pub fn enable_io(&mut self) -> &mut Builder

Enables the I/O driver.

Doing this enables using net, process, signal, and some I/O types on the runtime.

Examples

use tokio::runtime;

let rt = runtime::Builder::new_multi_thread()
    .enable_io()
    .build()
    .unwrap();

pub fn max_io_events_per_tick(&mut self, capacity: usize) -> &mut Builder

Enables the I/O driver and configures the max number of events to be processed per tick.

Examples

use tokio::runtime;

let rt = runtime::Builder::new_current_thread()
    .enable_io()
    .max_io_events_per_tick(1024)
    .build()
    .unwrap();

impl Builder

pub fn enable_time(&mut self) -> &mut Builder

Enables the time driver.

Doing this enables using tokio::time on the runtime.

Examples

use tokio::runtime;

let rt = runtime::Builder::new_multi_thread()
    .enable_time()
    .build()
    .unwrap();