headless_lms_chatbot::search_filter
Enum SearchFilter

pub enum SearchFilter {

    Eq(String, SearchFilterValue),
    Ne(String, SearchFilterValue),
    Gt(String, SearchFilterValue),
    Lt(String, SearchFilterValue),
    Ge(String, SearchFilterValue),
    Le(String, SearchFilterValue),
    In(String, Vec<String>),
    InWithDelimiter(String, Vec<String>, String),
    And(Box<SearchFilter>, Box<SearchFilter>),
    Or(Box<SearchFilter>, Box<SearchFilter>),
    Not(Box<SearchFilter>),
    Parentheses(Box<SearchFilter>),
    Raw(String),
    Field(String),
    Any(String, Option<Box<SearchFilter>>),
    All(String, Box<SearchFilter>),

}

A composable OData boolean expression.

Supports comparison operators, logical combinators, search.in, and raw sub-expressions.

Variants

Eq(String, SearchFilterValue)

Ne(String, SearchFilterValue)

Gt(String, SearchFilterValue)

Lt(String, SearchFilterValue)

Ge(String, SearchFilterValue)

Le(String, SearchFilterValue)

In(String, Vec<String>)

InWithDelimiter(String, Vec<String>, String)

search.in(field, 'v1|v2|...', '|') with custom delimiter

And(Box<SearchFilter>, Box<SearchFilter>)

Or(Box<SearchFilter>, Box<SearchFilter>)

Not(Box<SearchFilter>)

Parentheses(Box<SearchFilter>)

Explicit parentheses to override precedence

Raw(String)

Field(String)

Boolean field expression (e.g., IsEnabled)

Any(String, Option<Box<SearchFilter>>)

Collection any() operator (e.g., Rooms/any() or Rooms/any(room: room/BaseRate lt 200))

All(String, Box<SearchFilter>)

Collection all() operator (e.g., Rooms/all(room: not room/SmokingAllowed))

Implementations

impl SearchFilter

pub fn eq<T: Into<SearchFilterValue>>( field: impl Into<String>, value: T, ) -> Self

field eq value

pub fn ne<T: Into<SearchFilterValue>>( field: impl Into<String>, value: T, ) -> Self

field ne value

pub fn gt<T: Into<SearchFilterValue>>( field: impl Into<String>, value: T, ) -> Self

field gt value

pub fn lt<T: Into<SearchFilterValue>>( field: impl Into<String>, value: T, ) -> Self

field lt value

pub fn ge<T: Into<SearchFilterValue>>( field: impl Into<String>, value: T, ) -> Self

field ge value

pub fn le<T: Into<SearchFilterValue>>( field: impl Into<String>, value: T, ) -> Self

field le value

pub fn search_in(field: impl Into<String>, values: Vec<String>) -> Self

search.in(field, 'v1,v2,...', ',')

pub fn search_in_with_delimiter( field: impl Into<String>, values: Vec<String>, delimiter: &str, ) -> Self

search.in(field, 'v1|v2|...', '|') with custom delimiter

pub fn and(self, other: SearchFilter) -> Self

expr1 and expr2

pub fn or(self, other: SearchFilter) -> Self

expr1 or expr2

pub fn not(self) -> Self

not (expr)

pub fn parentheses(self) -> Self

Wrap expression in explicit parentheses to override precedence

pub fn raw(expr: impl Into<String>) -> Self

Insert an arbitrary OData sub-expression (e.g. geo.distance(...) or search.ismatchscoring(...))

pub fn field(field_name: impl Into<String>) -> Self

Boolean field expression (e.g., IsEnabled)

pub fn any(collection_path: impl Into<String>) -> Self

Collection any() operator without filter (e.g., Rooms/any())

pub fn any_with_filter( collection_path: impl Into<String>, filter: SearchFilter, ) -> Self

Collection any() operator with filter (e.g., Rooms/any(room: room/BaseRate lt 200))

pub fn all(collection_path: impl Into<String>, filter: SearchFilter) -> Self

Collection all() operator (e.g., Rooms/all(room: not room/SmokingAllowed))

pub fn to_odata(&self) -> Result<String, SearchFilterError>

Serialize this filter into a complete OData $filter string.

Errors

Returns SearchFilterError::ListNotSupportedInOData if any filter value contains a List variant that cannot be serialized to OData.

fn format_operand(filter: &SearchFilter) -> Result<String, SearchFilterError>

Helper function to format a filter operand, adding parentheses only when necessary. Simple expressions (comparisons, raw, field, etc.) don’t need parentheses, but complex logical expressions do.

fn to_odata_internal(&self) -> Result<String, SearchFilterError>