headless_lms_chatbot

Module azure_chatbot

Structs

ChatResponse

Choice

Citation

ContentFilter

ContentFilterResults

DataSource

DataSourceParameters

DataSourceParametersAuthentication

Delta

DeltaContext

EmbeddingDependency

FieldsMapping

GuardedStream 🔒

Custom stream that encapsulates both the response stream and the cancellation guard. Makes sure that the guard is always dropped when the stream is dropped.

LLMRequest

NonThinkingParams

RequestCancelledGuard 🔒

ResponseChunk

ThinkingParams

Enums

LLMRequestParams

Constants

CONTENT_FIELD_SEPARATOR 🔒

Functions

send_chat_request_and_parse_stream