Module headless_lms_chatbot::azure_chatbot

Structs

ApiChatMessage

The format accepted by the API.

ChatRequest

ChatResponse

Choice

ContentFilter

ContentFilterResults

DataSource

DataSourceParameters

DataSourceParametersAuthentication

Delta

EmbeddingDependency

FieldsMapping

GuardedStream 🔒

Custom stream that encapsulates both the response stream and the cancellation guard. Makes sure that the guard is always dropped when the stream is dropped.

RequestCancelledGuard 🔒

ResponseChunk

Constants

CHATBOT_AZURE_API_VERSION

Functions

build_headers 🔒

Build the headers for the API request.

estimate_tokens

Estimate the number of tokens in a given text. We use this for example to estimate the expense of a chat request. The result is not accurate but it is cheap to calculate.

send_chat_request_and_parse_stream