headless_lms_chatbot

Module chatbot_tools

Modules

course_progress

Structs

AzureLLMToolDefinition

A tool definition that is formatted for Azure. Defines a tool (function) that the LLM can call.

LLMTool

Content of an AzureLLMToolDefinition

LLMToolParamProperties

LLMToolParams

Parameters that a chatbot tool accepts in an AzureLLMToolDefinition

ToolProperties

Enums

LLMToolParamType

LLMToolType

Traits

ChatbotTool

Functions

get_chatbot_tool

Create a chatbot tool with LLM-provided arguments by matching the tool call made by the LLM. User context and db connection are needed for some tools.

get_chatbot_tool_definitions

Get a vec of AzureLLMToolDefinitions for all available chatbot tools